requirements.txt
# Core dependencies
selenium>=4.25.0
webdriver-manager>=4.0.2
beautifulsoup4>=4.12.3
requests>=2.32.3
urllib3>=2.2.3
httpx>=0.27.0

# Web Frameworks
flask>=3.1.0                    # Original Flask API
fastapi>=0.115.6                # Modern FastAPI Web UI
uvicorn[standard]>=0.34.0       # ASGI server for FastAPI
jinja2>=3.1.4                   # Templating engine
python-multipart>=0.0.20        # For form data handling

# Logging
loguru>=0.7.3

# Configuration
python-dotenv>=1.0.1

# HTML/XML parsing
lxml>=5.3.0

# Data validation
pydantic>=2.10.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
